Added error reporting to FilamentServiceProvider.
In app/Providers/FilamentServiceProvider.php, a new assignment to Page::$reportValidationErrorUsing was introduced. This provides a way to handle notifications for validation errors. Now, every time a ValidationException occurs, a notification will be created with the exception message and sent with a level of danger. This additional feature will significantly improve error management in our app.

Signed-off-by: Lloric Mayuga Garcia <lgarcia.halcyondigital@gmail.com>