"Add unique constraint and batchSize parameters to Excel import
This commit introduces the ability to specify a 'uniqueBy' field and batch size for Excel file imports. This is done by adding new parameters to the DefaultImport class and the ImportAction class which handles importing. These additions allow for efficient and accurate import of large Excel files, specifically ensuring uniqueness of entries based on the 'uniqueBy' parameter and controlling the size of data chunks with the 'batchSize' parameter."

Signed-off-by: Lloric Mayuga Garcia <lgarcia.halcyondigital@gmail.com>